+++ ui/package.json
@@ -10,6 +10,7 @@
     "react-dom": "^16.4.2",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
+    "socket.io": "^2.1.1",
     "styled-components": "^3.4.5"
   },
   "scripts": {
